requirements.txt
# 가상 환경 라이브러리 버전

streamlit==1.51.0
openai==2.8.1
python-dotenv==1.0.0
pandas==2.3.3
numpy==2.3.5
plotly==6.5.0